+++ src/lcm/grid_helpers.py
